add lqueue implementation, update matrix and queue types to use i32, and enhance readme for clarity

+++ include/queue.h
@@ -3,19 +3,19 @@
 #include "funcs.h"
 
 typedef struct Queue {
-  int *arr;
-  int size;
-  int capacity;
+  i32 *arr;
+  usize size;
+  usize capacity;
 
-  int front;
-  int back;
+  i32 front;
+  i32 back;
 } Queue;
 
 Queue init_queue(void);
 void print_queue(Queue *queue);
 void destroy_queue(Queue *queue);
 
-void enqeue_queue(Queue *queue, int n);
+void enqeue_queue(Queue *queue, i32 n);
 i32 dequeue_queue(Queue *queue);
 i32 peek_queue(Queue *queue);
