+++ src/bits.c
@@ -22,7 +22,7 @@ usize set_bit(usize num, size_t pos) {
   return num | mask;
 }
 usize set_bits(usize num, usize l, usize r) {
-  int mask = 0;
+  i32 mask = 0;
   for (size_t i = l; i < r; i++) {
     mask = mask | (1 << i);
   }
@@ -31,11 +31,11 @@ usize set_bits(usize num, usize l, usize r) {
 }
 
 usize clear_bit(usize num, size_t pos) {
-  int mask = ~(1 << pos);
+  i32 mask = ~(1 << pos);
   return mask & num;
 }
-usize clear_bits(usize num, int l, int r) {
-  int mask = ~0;
+usize clear_bits(usize num, usize l, usize r) {
+  i32 mask = ~0;
   for (size_t i = l; i < r; i++) {
     mask = mask | ~(1 << i);
   }
@@ -44,11 +44,11 @@ usize clear_bits(usize num, int l, int r) {
 }
 
 usize toggle_bit(usize num, size_t pos) {
-  int mask = 1 << pos;
+  i32 mask = 1 << pos;
   return num ^ mask;
 }
-usize toggle_bits(usize num, int l, int r) {
-  int mask = 0;
+usize toggle_bits(usize num, usize l, usize r) {
+  i32 mask = 0;
   for (size_t i = l; i < r; i++) {
     mask = mask | (1 << i);
   }
@@ -56,8 +56,8 @@ usize toggle_bits(usize num, int l, int r) {
   return num ^ mask;
 }
 
-int count_bits(usize num) {
-  int count = 0;
+i32 count_bits(usize num) {
+  i32 count = 0;
   while (num) {
     count += num & 1;
     num >>= 1;
@@ -66,7 +66,7 @@ int count_bits(usize num) {
 }
 
 void print_binary(usize num) {
-  int num_bits = sizeof(int) * __CHAR_BIT__;
+  i32 num_bits = sizeof(i32) * __CHAR_BIT__;
 
   usize mask = 1 << (num_bits - 1);
